12.02 15:08 commit

+++ src/main/java/org/swaphome/mapper/BoardAttachMapper.java
@@ -10,5 +10,5 @@ public interface BoardAttachMapper {
 	
 	public void delete(String uuid);
 	
-	public List<BoardAttachVO> findByBno(Long bno);
+	public List<BoardAttachVO> findByBno(int bno);
 }
